refactor(space): extract SpaceHeader and SpaceProjectList components, clean up space.tsx fix(spaces-card): make cards equal height with h-full + line-clamp-2 fix(landing): redirect authenticated users to /app, fix registration path fix(dropdown): raise z-index to prevent SearchBar overlap
